requirements.txt
pytz==2020.1
fastjsonschema==2.14.4
mock==4.0.2
boto3==1.14.33
json-logic-qubit==0.9.1
Django==3.0.9
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.12.1
django-redis==4.12.1
django-rq==2.3.2
djangorestframework-api-key==2.0.0
psycopg2-binary==2.8.5
pydash==4.8.0
gast==0.4.0
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
